Added -V flag to asVerify, and fixed a segmentation fault
Previously, if you ran `asVerify` with no filename, i.e. any of the
following:
```
$ asVerify
$ asVerify -v
$ asVerify -
```
you would get a segmentation fault due to the `strcpy` reaching beyond
the length of `argv`. This fixes this issue, as well as adding a `-V`
(for version) flag allowing one to test the existence of this utility.