Fix segmentation fault when trying to save invalid long strings
This can occur if you have something like the following:
```
record(waveform, "foo") {
  field(FTVL, "CHAR")
  field(NELM, "10")
  info(autosaveFields, "VAL VAL$")
}
```
The `VAL$` field will cause a segmentation fault when it tries to
save it to disk. The reason this seems to happen is that on one hand,
`VAL$` is regarded by autosave as a valid field (it is a field, but
with a `$` at the end). On the other hand, when connecting a monitor,
the connection fails as the `VAL` field is not of the correcct type
as defined in `dbChannelCreate` from EPICS base.

This means that autosave will try to save the date from an unconnected
array, with uninitialised pArray, causing a segmentation fault.